requirements.txt
torch
torchvision
ultralytics
roboflow
python-dotenv